+++ src/main/java/com/project/baedalsodae/store/dto/response/store/StorePageResponse.java
@@ -1,7 +1,6 @@
 package com.project.baedalsodae.store.dto.response.store;
 
 import com.project.baedalsodae.store.entity.Store;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
@@ -15,7 +14,7 @@ public class StorePageResponse {
     private String storeCategoryName;
 
     private boolean hasNext;
-    private BigDecimal storeCount;
+    private int storeCount;
     private UUID lastCursorId;
     private List<StoreSummaryResponse> stores;
 
@@ -24,7 +23,7 @@ public class StorePageResponse {
                 .storeCategoryId(categoryId)
                 .storeCategoryName(categoryName)
                 .hasNext(slice.hasNext())
-                .storeCount(BigDecimal.valueOf(slice.getNumberOfElements()))
+                .storeCount(slice.getNumberOfElements())
                 .lastCursorId(getLastCursorId(slice))
                 .stores(StoreSummaryResponse.fromList(slice.getContent()))
                 .build();
