+++ src/main/java/com/project/baedalsodae/store/dto/response/storeCategory/StoreCategoryListResponse.java
@@ -1,6 +1,5 @@
 package com.project.baedalsodae.store.dto.response.storeCategory;
 
-import java.math.BigDecimal;
 import java.util.List;
 import lombok.*;
 
@@ -8,13 +7,13 @@ import lombok.*;
 @Builder
 public class StoreCategoryListResponse {
     private List<StoreCategoryResponse> storeCategoryList;
-    private BigDecimal totalCount;
+    private int totalCount;
 
     public static StoreCategoryListResponse fromList(
             List<StoreCategoryResponse> storeCategoryList) {
         return StoreCategoryListResponse.builder()
                 .storeCategoryList(storeCategoryList)
-                .totalCount(BigDecimal.valueOf(storeCategoryList.size()))
+                .totalCount(storeCategoryList.size())
                 .build();
     }
 }
